requirements.txt
streamlit
PyYAML
google-generativeai
